+++ commons/src/test/java/com/karuslabs/commons/command/dispatcher/ExceptionsTest.java
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package com.karuslabs.commons.command;
+package com.karuslabs.commons.command.dispatcher;
 
 import com.karuslabs.commons.command.dispatcher.Exceptions;
 import com.mojang.brigadier.LiteralMessage;